requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
python-chess>=1.9.0
tqdm>=4.62.0
svgwrite>=1.4.0
